Separating HuggingFace Dataset Preprocessing
- Moveing process_split, load_and_split_hf_dataset, and related HuggingFace preprocessing functions from fasta.py to hf_dataset_to_fasta.py in utilities
- Improving code organization by separating concerns between general FASTA utilities and HuggingFace dataset processing
